myrobotictrader-mcp/requirements.txt
# MyRoboticTrader MCP Server Dependencies
# ========================================

# Core MCP server framework
mcp>=1.0.0

# HTTP client for API calls (NewsAPI, Discord, CoinMarketCap)
httpx>=0.27.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Google Sheets integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Anthropic API for content generation (optional - uncomment when ready)
# anthropic>=0.18.0

# Development tools (optional)
# pytest>=7.4.0
# black>=23.0.0
